bot notify

+++ bot.py
@@ -1,10 +1,10 @@
 import telebot
 
 import config
+from handler.news import news
 from handler.schedule import add_schedule, view_schedule, delete_schedule_all, delete_schedule_id
 
 bot = telebot.TeleBot(config.token_tele)
-
 status = None
 day = None
 day_int = None
@@ -14,18 +14,26 @@ result = []
 data_del = []
 i = 1
 j = 0
+admin = 1457896502
 
 
 @bot.message_handler(commands=['start', 'help'])
 def send_welcome(message):
-    bot.reply_to(message, "Hello, day la chuong trinh bot_telegram thong bao thoi khoa bieu.")
-    bot.send_message(message.chat.id, text="/add_schedule\n/view_schedule")
+    bot.reply_to(message, text="Tôi là ai??\nTôi đang ở đâu thế này??\nview_schedule\nnews")
+    if message.from_user.id == admin:
+        bot.send_message(message.chat.id, text="/add_schedule")
+        return
 
 
 @bot.message_handler(commands=["add_schedule"])
 def add_sche(message):
-    text = "Chon ngay.\n/mon\n/tue\n/wed\n/thu\n/fri"
-    bot.reply_to(message, text=text)
+    if message.from_user.id == admin:
+        text = "Chon ngay.\n/mon\n/tue\n/wed\n/thu\n/fri"
+        bot.reply_to(message, text=text)
+        return
+    else:
+        bot.reply_to(message, text="Cu chỉ được xem thôi..")
+        return
 
 
 @bot.message_handler(commands=["view_schedule"])
@@ -44,25 +52,26 @@ def view_sche(message):
             data_del.append(item_del)
         j = 0
         bot.send_message(message.chat.id, text=view_str)
-        bot.send_message(message.chat.id,
-                         text="/add_schedule de them mon hoc.\n/delete_id de xoa mon hoc theo id\n/delete_all de xoa "
-                              "tat ca cac mon hoc")
         return
-    bot.send_message(message.chat.id, text="Thoi khoa bieu cua ban khong co du lieu")
-    bot.send_message(message.chat.id, text="/add_schedule de them cac mon hoc")
+    bot.send_message(message.chat.id, text="Thời khóa biểu không có dữ liệu.")
+    bot.send_message(message.chat.id, text="Đợi chủ nhân 'Oh Fuoc' của tôi cho tôi ăn đã. 🤤")
     return
 
 
 @bot.message_handler(commands=["delete_all", "delete_id"])
 def delete_subject(message):
     global status
-    if message.text == "/delete_all":
-        delete_schedule_all()
-        bot.send_message(message.chat.id, text="Ban da xoa thanh cong cac mon hoc")
-        return
-    elif message.text == "/delete_id":
-        bot.send_message(message.chat.id, text="nhap cac id muon xoa cach nhau bang dau ','\nVi du: 1, 2, 3, 4 ...")
-        status = "delete_id"
+    if message.from_user.id == admin:
+        if message.text == "/delete_all":
+            delete_schedule_all()
+            bot.send_message(admin, text="Ban da xoa thanh cong cac mon hoc")
+            return
+        elif message.text == "/delete_id":
+            bot.send_message(admin, text="nhap cac id muon xoa cach nhau bang dau ','\nVi du: 1, 2, 3, 4 ...")
+            status = "delete_id"
+            return
+    else:
+        bot.reply_to(message, text="??? Định làm gì cơ..?")
         return
 
 
@@ -72,20 +81,30 @@ def add_schedule_with_day(message):
     global day
     global day_int
     global count
-    bot.reply_to(message, text=f"You chosen {message.text}.")
-    day = str(message.text).strip("/").title()
-    if day == "Mon":
-        day_int = 2
-    elif day == "Tue":
-        day_int = 3
-    elif day == "Wed":
-        day_int = 4
-    elif day == "Thu":
-        day_int = 5
-    elif day == "Fri":
-        day_int = 6
-    bot.send_message(message.chat.id, "Nhap so mon ban muon nhap")
-    status = "vao viec"
+    if message.from_user.id == admin:
+        bot.reply_to(message, text=f"You chosen {message.text}.")
+        day = str(message.text).strip("/").title()
+        if day == "Mon":
+            day_int = 2
+        elif day == "Tue":
+            day_int = 3
+        elif day == "Wed":
+            day_int = 4
+        elif day == "Thu":
+            day_int = 5
+        elif day == "Fri":
+            day_int = 6
+        bot.send_message(admin, "Nhap so mon ban muon nhap")
+        status = "vao viec"
+    else:
+        bot.reply_to(message, text="??? Định làm cái quái gì nữa...")
+        return
+
+
+@bot.message_handler(commands=["news"])
+def view_news(message):
+    news(message=message, bot=bot)
+    return
 
 
 @bot.message_handler(func=lambda msg: True)
@@ -96,69 +115,73 @@ def reply(message):
     global result
     global i
 
-    if status == "vao viec":
-        try:
-            if isinstance(int(message.text), int):
-                count = int(message.text)
-            status = "add_subject"
-        except ValueError:
-            bot.reply_to(message, "Ban nhap sai lua chon")
-            return
-    if status == "add_subject":
-        while count > 0:
-            if "subject" not in par_item:
-                try:
-                    if isinstance(int(message.text), int):
-                        bot.send_message(message.chat.id, f"Moi ban nhap subjects {i}: ")
+    if message.chat.id == admin:
+        if status == "vao viec":
+            try:
+                if isinstance(int(message.text), int):
+                    count = int(message.text)
+                status = "add_subject"
+            except ValueError:
+                bot.reply_to(message, "Ban nhap sai lua chon")
+                return
+        if status == "add_subject":
+            while count > 0:
+                if "subject" not in par_item:
+                    try:
+                        if isinstance(int(message.text), int):
+                            bot.send_message(admin, f"Moi ban nhap subjects {i}: ")
+                            return
+                    except ValueError:
+                        par_item["subject"] = message.text
+                        bot.send_message(admin, f"Moi ban nhap lesson {i}: ")
                         return
-                except ValueError:
-                    par_item["subject"] = message.text
-                    bot.send_message(message.chat.id, f"Moi ban nhap lesson {i}: ")
+
+                if "lesson" not in par_item:
+                    par_item["lesson"] = message.text
+                    bot.send_message(admin, f"Moi ban nhap room {i}: ")
                     return
 
-            if "lesson" not in par_item:
-                par_item["lesson"] = message.text
-                bot.send_message(message.chat.id, f"Moi ban nhap room {i}: ")
-                return
+                if "room" not in par_item:
+                    par_item["room"] = message.text
+                    par_item["day"] = day
+                    par_item["day_int"] = day_int
+                    message.text = 1
+                    result.append(par_item)
+                    par_item = {}
 
-            if "room" not in par_item:
-                par_item["room"] = message.text
-                par_item["day"] = day
-                par_item["day_int"] = day_int
-                message.text = 1
-                result.append(par_item)
-                par_item = {}
+                count -= 1
+                i += 1
+            i = 1
+            add_schedule(data=result)
+            bot.send_message(admin, text=f"Ban da nhap xong {result}")
+            par_item = {}
+            result = []
+            status = None
+            return
 
-            count -= 1
-            i += 1
-        i = 1
-        add_schedule(data=result)
-        bot.send_message(message.chat.id, text=f"Ban da nhap xong {result}")
-        par_item = {}
-        result = []
-        status = None
+        elif status == "delete_id":
+            global data_del
+            solve_text = str(message.text).split(",")
+            new_data = []
+            for item in solve_text:
+                try:
+                    isinstance(int(item), int)
+                    new_data.append(int(item))
+                except ValueError:
+                    bot.reply_to(message, text="Hinh nhu ban nhap sai cu phap")
+                    bot.send_message(admin, text="/delete_id de thuc hien lai")
+                    return
+
+            delete_schedule_id(new_data, data_del=data_del)
+            data_del = []
+            bot.send_message(admin, text="Ban da xoa thanh cong cac mon hoc theo id")
+            status = None
+            return
+    print(message.chat.id)
+    if message.from_user.id == admin:
+        bot.reply_to(message, text="Dạ thưa chủ nhân, em không biết ngài cần gì ạ.. 🥺")
         return
-
-    elif status == "delete_id":
-        global data_del
-        solve_text = str(message.text).split(",")
-        new_data = []
-        for item in solve_text:
-            try:
-                isinstance(int(item), int)
-                new_data.append(int(item))
-            except ValueError:
-                bot.reply_to(message, text="Hinh nhu ban nhap sai cu phap")
-                bot.send_message(message.chat.id, text="/delete_id de thuc hien lai")
-                return
-
-        delete_schedule_id(new_data, data_del=data_del)
-        data_del = []
-        bot.send_message(message.chat.id, text="Ban da xoa thanh cong cac mon hoc theo id")
-        status = None
-        return
-
-    bot.send_message(message.chat.id, "Toi khong hieu ban muon gi")
+    bot.send_message(message.chat.id, "Đừng làm những điều vớ vẩn, vì bạn không phải là chủ nhân của tôi. 😌")
 
 
 bot.infinity_polling()
